remove todo comments
Also removed pending tests that should be added when the functionality they are related to is added to the code base.  All removed items were either already addressed or have outstanding tickets.  There were a couple of TODOs that were not removed, but they are marked with the issue number that is addressing that TODO.