Removing memoized object as query is stateless
There was not a benefit to memoizing the AncestorChecker object. So I
opted to remove the memoization and favor a module function.

I also did some parameter renaming to reflect that the parameter names
switched meaning across the call into the AncestorChecker.

And finally added a missing spec for potential deep ancestry.