Removing DeepMemberIterator
Based on the previous commit, the DeepMemberIterator is not needed.
I checked through the two organizations (i.e. projectyhdra and
projecthydra-labs) for any additional uses of DeepMemberIterator, and
found none.

Perhaps this should be deprecated instead. I don't know if the class is
part of the public api or not.

(If it isn't perhaps we should mark the constant as a private constant
or document it as @api private)?