Renaming opaque `.call` method
When I was orienting to the code, it took a bit to understand the
ancestor? method ordering and purpose. As I was extracting a module
function, I opted to go with `.call`. However, when I went back to the
method, I had to once again reorient to what was going on.

So I opted for the new method name `.former_is_ancestor_of_latter?`.
This resulted in a change in parameter order. To preserve parameter
order, I considered the method `.latter_is_ancestor_of_former?` but
that was more difficult to parse when reading.