Test demonstration OO-extensibility failure of #generate_solr_document
when used with passed block.

Note the same test expectation (that the yielded object is the same as
the object returned), passes with the base case, but fails once the
class is extended and the method overridden.  This is not a pathological
example, since it is exactly how we have been telling consumers to
extend this class (among others). The problem has to do with the
"timing" of the yield, happening at the end of the base class method,
not at the end of the overriding method.