Forces Circle to report build status to Coveralls
When using parallel builds, Coveralls does not always know when a build
is completed. Circle's deploy step only happen if the other parallel
nodes finish successfully. Therefore, we can use the deploy step to tell
Coveralls, once per build, that a build has completed successfully.

In support of this, the Coveralls repository token has been added to
CircleCI for the hydra-pcdm project.